refactor(cart): redesign AddItemToShoppingCartUseCase and ShoppingCartGateway
- Rename ShoppingCartRepository to ShoppingCartGateway
- Replace getActiveCartByUserId() with getLastCreatedCart()
- Remove separate update() method from ShoppingCartGateway
- Extract getOrCreateCart() private method
- Replace inline null check with NullValidator
- Replace ProductSummaryOut with ProductOut in ShoppingCartItemOut
- Update ShoppingCartItemMapper to use ProductMapper
- Remove unused ProductSummaryOutMapper from ProductService
