+++ core/application/src/main/java/com/arka/gateway/repository/ShoppingCartGateway.java
@@ -1,18 +1,15 @@
 package com.arka.gateway.repository;
 
-import com.arka.dto.out.ShoppingCartOut;
 import com.arka.model.cart.ShoppingCart;
 
 import java.util.List;
 import java.util.Optional;
 
-public interface ShoppingCartRepository {
+public interface ShoppingCartGateway {
 
     ShoppingCart save(ShoppingCart shoppingCart);
 
     List<ShoppingCart> getAllAbandonedCarts();
 
-    Optional<ShoppingCart> getActiveCartByUserId(Long userId);
-
-    ShoppingCart update(ShoppingCart shoppingCart);
+    Optional<ShoppingCart> getLastCreatedCart(Long userId);
 }
