+++ core/application/src/main/java/com/arka/mapper/ShoppingCartItemMapper.java
@@ -4,7 +4,7 @@ import com.arka.dto.out.ShoppingCartItemOut;
 import com.arka.model.cart.ShoppingCartItem;
 import org.mapstruct.Mapper;
 
-@Mapper(uses = ProductSummaryOutMapper.class)
+@Mapper(uses = ProductMapper.class)
 public interface ShoppingCartItemMapper {
 
     ShoppingCartItemOut toOutDTO(ShoppingCartItem domain);
