+++ core/application/src/main/java/com/arka/usecase/cart/AddItemToShoppingCartUseCase.java
@@ -1,22 +1,21 @@
-package com.arka.usecase;
+package com.arka.usecase.cart;
 
 import com.arka.dto.in.AddItemShoppingCartIn;
 import com.arka.dto.out.ShoppingCartOut;
-import com.arka.gateway.repository.ShoppingCartRepository;
+import com.arka.gateway.repository.ShoppingCartGateway;
 import com.arka.mapper.ShoppingCartMapper;
 import com.arka.mapper.ShoppingCartMapperImpl;
 import com.arka.model.cart.ShoppingCart;
 import com.arka.model.product.Product;
 import com.arka.service.ProductService;
 import com.arka.service.WarehouseInventoryService;
+import com.arka.util.NullValidator;
 import lombok.RequiredArgsConstructor;
-
-import java.util.Optional;
 
 @RequiredArgsConstructor
 public class AddItemToShoppingCartUseCase {
 
-    private final ShoppingCartRepository cartRepository;
+    private final ShoppingCartGateway cartGateway;
 
     private final ProductService productService;
     private final WarehouseInventoryService inventoryService;
@@ -24,38 +23,25 @@ public class AddItemToShoppingCartUseCase {
     private final ShoppingCartMapper mapper =
             new ShoppingCartMapperImpl();
 
-    public ShoppingCartOut execute(AddItemShoppingCartIn request){
+    public ShoppingCartOut execute(AddItemShoppingCartIn input){
 
-        if(request == null)
-            throw new IllegalArgumentException("Request can't be null");
+        NullValidator.validate(input, "input");
 
         inventoryService.validateGeneralStockAvailability(
-                request.productId(), request.quantity()
-        );
+                input.productId(), input.quantity());
 
         Product foundProduct =
-                productService.findById(request.productId());
+                productService.findById(input.productId());
 
-        Optional<ShoppingCart> activeCart = cartRepository
-                .getActiveCartByUserId(request.userId())
-                .map(cart -> {
+        ShoppingCart cart = getOrCreateCart(input.userId());
 
-                    cart.addItem(foundProduct, request.quantity());
-                    return cart;
+        cart.addItem(foundProduct, input.quantity());
 
-                });
+        return mapper.toOutDto(cartGateway.save(cart));
+    }
 
-        if(activeCart.isPresent())
-            return mapper.toOutDto(cartRepository.update(activeCart.get()));
-
-         else {
-
-            ShoppingCart newCart = new ShoppingCart(request.userId());
-            newCart.addItem(foundProduct, request.quantity());
-
-            return mapper.toOutDto(cartRepository.save(newCart));
-        }
-
-
+    private ShoppingCart getOrCreateCart(Long userId){
+        return cartGateway.getLastCreatedCart(userId)
+                .orElseGet(() -> ShoppingCart.create(userId));
     }
 }
